+++ examples/03-dev/demo/project/dev/html/test-3/index.js
@@ -2,8 +2,7 @@
 module.exports = {
     config: {
         packCssSeparately: !0,
-        splitJs: !0,
-        splitJsMap: {
+        splitJs: {
             vendor: ['jquery', 'underscore'],
             vendor2: ['react', 'react-dom'],
             common: ['alias/base', 'alias/common']