+++ examples/03-dev/demo/project/dev/html/test-4/index.js
@@ -3,8 +3,7 @@ module.exports = {
     config: {
         revisionFiles: !1,
         packCssSeparately: !0,
-        splitJs: !0,
-        splitJsMap: {
+        splitJs: {
             vendor: ['jquery', 'underscore', 'react', 'react-dom']
         }
     }