+++ project_config/defaults/base.js
@@ -44,8 +44,6 @@ var defaults = {
     cssAutoPrefix: false,
     // whether to pack css separately into a single css file
     packCssSeparately: false,
-    // whether to split one big js file into many smaller js files
-    splitJs: false,
     // whether to ignore files under node_modules directory when transform es6 to es5
     ignoreNodeModules: true,
     // indicates which files to load