+++ project_files/config/lila.clean.config.js
@@ -1,8 +1,0 @@
-
-"use strict";
-
-module.exports = {
-    // hash code length
-    hashLength: 8
-};
-