+++ project_files/demo/base.js
@@ -1,4 +1,2 @@
 
-require('{{cssPath}}.css');
-
 // start here